Dix deprecation warning about converting bool column to object
The warning is:
FutureWarning: Setting an item of incompatible dtype is deprecated and will raise an error in a future version of pandas. Value 'nan' has dtype incompatible with bool, please explicitly cast to a compatible dtype first.
    new_column[missing_vals] = np.nan

And it's happening because new_column has type bool but needs to be coreced to object to support nan values.